Allow users to zoom in sounding diagrams
Sounding diagrams are now reactive and can show either the boundary layer or the whole troposphere.
